requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pytest>=8.0.0
httpx>=0.27.0

pydantic-settings
google-genai
boto3
sqlalchemy
asyncpg
alembic
python-dotenv
psycopg2-binary
dotenv
